requirements.txt
Flask==0.12.2
Flask-Login==0.4.1
mysql-connector-python-rf==2.2.2
Faker==0.8.14
